paginacion de categorias, y renombre de etiquetas y variables

+++ src/public/js/categories.js
@@ -28,26 +28,39 @@ $(document).ready(() => {
         resetForm();
     });
 
+    // Variables para controlar la paginación de libros en general
+    let currentPageCategories = 1;
+    let totalPagesCategories = 1;
 
-    const getCategories = async () => {
-        const categories = await categoryServices.getCategories();
-        const tbody = $("#tbody");
 
-        tbody.html(
-            categories.map((category) => {
-                return `
-                <tr>
-                    <td>${category.name}</td>
-                    <td>
-                        <button class="edit" data-id="${category._id}">Editar</button>
-                        <button class="delete" id="delete" data-id="${category._id}">Eliminar</button>
-                    </td>
-                </tr>
-                `;
-            })
-        );
+
+
+    const getCategories = async (page = 1) => {
+        try {
+            const categories = await categoryServices.getCategories(page);
+            totalPagesCategories = categories.totalPages;
+            const tbody = $("#tbody");
+            tbody.html(
+                categories.docs.map((category) => {
+                    return `
+            <tr>
+                <td>${category?.name}</td>
+                <td>
+                    <button class="edit" data-id="${category._id}">Editar</button>
+                    <button class="delete" id="delete" data-id="${category._id}">Eliminar</button>
+                </td>
+            </tr>
+            `;
+                })
+            );
+
+            $("#pageCat").val(page);
+            $("#inicio").text(`${currentPageCategories}`);
+            $("#final").text(`${totalPagesCategories}`);
+        } catch (error) {
+            alert(error);
+        }
     };
-    getCategories();
 
     const createandeditCategories = async (e) => {
         e.preventDefault();
@@ -55,43 +68,71 @@ $(document).ready(() => {
             name: $("#name").val(),
         };
 
-        if (editingCategoryId) {
-            await categoryServices.updateCategory(editingCategoryId, categories);
-            resetForm();
-            editingCategoryId = null;
-            $("#edit").text("Guardar").attr("id", "create");
-            modal.close();
-        } else {
-            await categoryServices.createCategory(categories);
-            modal.close();
+        try {
+            if (editingCategoryId) {
+                await categoryServices.updateCategory(editingCategoryId, categories);
+                resetForm();
+                editingCategoryId = null;
+                $("#edit").text("Guardar").attr("id", "create");
+                modal.close();
+            } else {
+                await categoryServices.createCategory(categories);
+                modal.close();
+            }
+
+        } catch (error) {
+            console.error(error);
         }
+
         resetForm();
-        getCategories(); // Refrescamos la lista actual de libros
+        getCategories(currentPageCategories); // Refrescamos la lista actual de libros
     };
 
     const deleteCategories = async (e) => {
         e.preventDefault();
         const id = $(e.target).data("id");
-        await categoryServices.deleteCategory(id);
-        getCategories(); // Refrescamos la lista actual de libros
+        const confirmation = confirm("¿Seguro que quieres borrar esta categoría?");
+        if (!confirmation) return;
+        try {
+            await categoryServices.deleteCategory(id);
+            getCategories(currentPageCategories); // Refrescamos la lista actual de libros
+        } catch (error) {
+            console.error(error);
+        }
     };
 
     const editingCategory = async (e) => {
         e.preventDefault();
         const categoriesId = $(e.target).data("id");
-        const categories = await categoryServices.getCategory(categoriesId);
-        console.log(categories);
 
-        $("#name").val(categories.name);
+        try {
+            const categories = await categoryServices.getCategory(categoriesId);
+            $("#name").val(categories.name);
+            editingCategoryId = categoriesId;
+            $("#create").text("Actualizar").attr("id", "edit");
+            modal.showModal();
+        } catch (error) {
+            console.error(error);
 
-        editingCategoryId = categoriesId;
-        console.log(editingCategoryId);
-
-        $("#create").text("Actualizar").attr("id", "edit");
-
-        modal.showModal();
+        }
     };
 
+
+    $("#pageprevCat").on("click", () => {
+        if (currentPageCategories > 1) {
+            currentPageCategories--;
+            getCategories(currentPageCategories);
+        }
+    });
+
+    $("#pageafterCat").on("click", () => {
+        if (currentPageCategories < totalPagesCategories) {
+            currentPageCategories++;
+            getCategories(currentPageCategories);
+        }
+    });
+
+    getCategories(currentPageCategories);
 
     $("#create").on("click", createandeditCategories);
     $(document).on("click", ".delete", deleteCategories);
